marketplace/requirements.txt
fastapi[all]
grpcio-tools
Jinja2
pytest